project18_fraud_detection_ai/requirements.txt
# Core ML Libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1

# Deep Learning
tensorflow==2.13.0
keras==2.13.1
torch==2.0.1

# Anomaly Detection
pyod==1.1.0

# API & Web
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-multipart==0.0.6

# Dashboard
streamlit==1.26.0
plotly==5.16.1
streamlit-autorefresh==0.0.1

# MLflow
mlflow==2.6.0

# Data Processing
imbalanced-learn==0.11.0

# Utilities
python-dotenv==1.0.0
loguru==0.7.0
requests==2.31.0
joblib==1.3.2

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Monitoring
prometheus-client==0.17.1
psutil==5.9.5
